move side panel components
