+++ client/src/components/map/SidePanel/About.js
