+++ client/src/components/map/SidePanel/About.scss
