+++ client/src/components/map/SidePanel/BufferSlider.js
