+++ client/src/components/map/SidePanel/BufferSlider.scss
