+++ client/src/components/map/SidePanel/GeocoderInput.scss
