+++ client/src/components/map/SidePanel/MarkerSelector.js
