+++ client/src/components/map/SidePanel/PDIIndicator.scss
