+++ client/src/components/map/SidePanel/PermitType.scss
