+++ client/src/components/map/SidePanel/PermitTypeText.js
