+++ client/src/components/map/SidePanel/SidePanel.scss
@@ -16,7 +16,7 @@
   justify-content: space-between;
   z-index: 3;
   width: 375px;
-  margin: 35px 0 20px 0;
+  margin: 40px 0 20px 0;
   //just need to set a max height to get overflow
   max-height: calc(100vh - 100px);
   overflow-y: scroll;
