+++ client/src/components/map/SidePanel/SiteDetails.js
