+++ client/src/components/map/SidePanel/SiteDetails.scss
